Refactor: Update import paths and add subcategory constants for my work and college work
