+++ backend/src/constants.js
@@ -4,3 +4,6 @@ export const options = {
     secure: true,
     sameSite: 'None'
 }
+
+export const myWorkSubcats = ["design", "logo_design", "mask_making", "book_cover", "other"];
+export const collegeWorkSubcats = ["model_making", "sand_art", "other"];
