+++ backend/src/controllers/myWork.controller.js
@@ -5,7 +5,7 @@ import {MyWork} from "../models/myWork.model.js";
 import { Image } from "../models/images.model.js";
 import { uploadOnCloudinary, deleteFromCloudinary, getPublicId } from "../utils/cloudinary.js";
 import mongoose from "mongoose";
-import { myWorkSubcats,collegeWorkSubcats } from "../../../const.js";
+import { myWorkSubcats,collegeWorkSubcats } from "../constants.js";
 
 const uploadMyWorkImage = asyncHandler(async (req, res) => {
     const { title, description } = req.body;
