+++ frontend/krishti/src/constant/constant.js
@@ -5,3 +5,6 @@ export function formatName(input) {
         .map((word) => word.charAt(0).toUpperCase() + word.slice(1).toLowerCase())
         .join(" ");
 }
+
+export const myWorkSubcats = ["design", "logo_design", "mask_making", "book_cover", "other"];
+export const collegeWorkSubcats = ["model_making", "sand_art", "other"];
